# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeHub - Deriv Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trade journal entries
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: result
          in: query
          required: false
          schema:
            type: string
            enum: [win, loss, breakeven]
      responses:
        "200":
          description: Trade entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Add a manual trade journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTradeBody"
      responses:
        "201":
          description: Trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        contractType:
          type: string
          description: e.g. CALL, PUT, RISE, FALL, DIGITS
        stake:
          type: number
        payout:
          type: number
        profit:
          type: number
        result:
          type: string
          enum: [win, loss, breakeven]
        botId:
          type: integer
          nullable: true
        notes:
          type: string
        tradedAt:
          type: string
          format: date-time
      required: [id, symbol, contractType, stake, payout, profit, result, tradedAt]
    CreateTradeBody:
      type: object
      properties:
        symbol:
          type: string
          minLength: 1
        contractType:
          type: string
          minLength: 1
        stake:
          type: number
        payout:
          type: number
        profit:
          type: number
        result:
          type: string
          enum: [win, loss, breakeven]
        botId:
          type: integer
          nullable: true
        notes:
          type: string
        tradedAt:
          type: string
          format: date-time
      required: [symbol, contractType, stake, payout, profit, result]
...